+++ src/components.d.ts
@@ -7,9 +7,10 @@ export {}
 /* prettier-ignore */
 declare module 'vue' {
   export interface GlobalComponents {
-    AnnotationTool: typeof import('./components/AnnotationTool.vue')['default']
-    ConnectionDialog: typeof import('./components/ConnectionDialog.vue')['default']
-    DrawToolbar: typeof import('./components/DrawToolbar.vue')['default']
+    AnnotationPopup: typeof import('./components/Tooltip/AnnotationPopup.vue')['default']
+    AnnotationTool: typeof import('./components/Tooltip/AnnotationTool.vue')['default']
+    ConnectionDialog: typeof import('./components/DrawToolbar/ConnectionDialog.vue')['default']
+    DrawToolbar: typeof import('./components/DrawToolbar/DrawToolbar.vue')['default']
     ElButton: typeof import('element-plus/es')['ElButton']
     ElButtonGroup: typeof import('element-plus/es')['ElButtonGroup']
     ElCard: typeof import('element-plus/es')['ElCard']
@@ -31,11 +32,11 @@ declare module 'vue' {
     ElSelect: typeof import('element-plus/es')['ElSelect']
     ElSwitch: typeof import('element-plus/es')['ElSwitch']
     ElTree: typeof import('element-plus/es')['ElTree']
-    ExternalResourceCard: typeof import('./components/ExternalResourceCard.vue')['default']
-    HelpModeDialog: typeof import('./components/HelpModeDialog.vue')['default']
-    ProvenancePopup: typeof import('./components/ProvenancePopup.vue')['default']
-    Tooltip: typeof import('./components/Tooltip.vue')['default']
-    TreeControls: typeof import('./components/TreeControls.vue')['default']
+    ExternalResourceCard: typeof import('./components/Tooltip/ExternalResourceCard.vue')['default']
+    HelpModeDialog: typeof import('./components/HelpModeDialog/HelpModeDialog.vue')['default']
+    ProvenancePopup: typeof import('./components/Tooltip/ProvenancePopup.vue')['default']
+    Tooltip: typeof import('./components/Tooltip/Tooltip.vue')['default']
+    TreeControls: typeof import('./components/TreeControls/TreeControls.vue')['default']
   }
   export interface ComponentCustomProperties {
     vLoading: typeof import('element-plus/es')['ElLoadingDirective']
