+++ src/components/HelpModeDialog/HelpModeDialog.vue
