+++ src/components/Tooltip/ExternalResourceCard.vue
@@ -21,7 +21,7 @@
 import { shallowRef } from "vue";
 import { Notebook as ElIconNotebook } from "@element-plus/icons-vue";
 
-import EventBus from "./EventBus";
+import EventBus from "../EventBus";
 
 export default {
   name: "ExternalResourceCard",
