+++ src/components/Tooltip/Tooltip.vue
@@ -1,7 +1,7 @@
 <template>
   <div class="tooltip-container" id="tooltip-container">
     <template v-if="annotationDisplay">
-      <annotation-tool
+      <annotation-popup
         :annotationEntry="annotationEntry"
         @annotation="$emit('annotation', $event)"
       />
